create area chart custom component
