+++ src-tauri/src/metrics.rs
@@ -27,7 +27,7 @@ struct Metrics {
 
 impl Metrics {
     fn memory(&mut self) -> Memory {
-        self.sys.refresh_all();
+        self.sys.refresh_memory();
         let unit = ByteUnit::GB;
         let free = kb_to_size(self.sys.free_memory(), &unit);
         let total = kb_to_size(self.sys.total_memory(), &unit);
